requirements.txt
certifi==2020.4.5.1
click==7.1.2
elasticsearch==7.7.1
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dotenv==0.13.0
python-json-logger==0.1.11
python-logstash==0.4.6
python3-logstash==0.4.80
urllib3==1.25.9
Werkzeug==1.0.1
